hima

+++ include/joseph_perf.h
@@ -80,7 +80,7 @@ extern "C" {
 #endif
 	void jperf_usage();
 
-	uint32_t getmicro();
+	unsigned long getmicro();
 	void jperf_start();
 	void jperf_stop();
 	unsigned long jperf_time();
